Added export to CSV in admin panel and attaching PDF to email to customer

+++ orders/models.py
@@ -30,9 +30,7 @@ class OrderItem(models.Model):
     product = models.ForeignKey(Product,
                                 related_name='order_items',
                                 on_delete=models.CASCADE)
-    price = models.DecimalField(max_digits=10,
-                                decimal_places=2,
-                                verbose_name='Цена')
+    price = models.PositiveIntegerField(verbose_name='Цена')
     quantity = models.PositiveIntegerField(default=1,
                                            verbose_name='Количество')
